+++ .eslintrc
@@ -4,6 +4,7 @@
     "browser": true,
     "es6": true
   },
+  "parser": "babel-eslint",
   "ecmaFeatures": {
     "modules": true,
     "jsx": true
@@ -120,7 +121,7 @@
       "no-wrap-func": 2,
       "no-underscore-dangle": 0,
       "one-var": [2, "never"],
-      "padded-blocks": [0],
+      "padded-blocks": [2, "never"],
       "semi-spacing": [2, {
         "before": false,
         "after": true
